Split SamacSys WRL regression tests
